neat: move token hashing functions to test utils
